Configuring eslint and prettier
Setting up ESLint and Prettier for this repo

At @webmasterkai's suggestion, we are mostly using the Airbnb JS Style guide
except tweaking it to use semicolons as little as possible.

I set this up using this script: https://github.com/paulolramos/eslint-prettier-airbnb-react in the hopes that it would make sure that eslint and prettier
would play nicely together. I think then manually edited the configuration
files to set semi to never/false.

I didn't run prettier or ESLint autofix over all the files in the repository
as part of this change. That could be done as separate commit if someone
wants to do it.

I also didn't setup any commit hooks or other enforcement mechanisms for
these policies. If these seem to be working well and we want to enforce
them more strongly later, we could add that. In the short term, it seems
like those could risk blocking us from pushing or editing the website and
that would be worse than having inconsistent style or formatting.